refator cart feature connection to frontend by using custom hook
